+++ lib/flutter_dropdown_button.dart
@@ -7,6 +7,7 @@ export 'src/overlay/dropdown_overlay_controller.dart';
 export 'src/presentation/item_presentation.dart';
 export 'src/search/dropdown_search_controller.dart';
 export 'src/config/text_dropdown_config.dart';
+export 'src/theme/dropdown_checkbox_theme.dart';
 export 'src/theme/dropdown_scroll_theme.dart';
 export 'src/theme/dropdown_style_theme.dart';
 export 'src/theme/dropdown_theme.dart';